+++ tests/bench/tests/binary-size.ts
@@ -1,3 +1,4 @@
+import { execSync } from "child_process";
 import * as fs from "fs/promises";
 import path from "path";
 
@@ -9,8 +10,12 @@ describe("Binary size", () => {
   const binarySize: BinarySize = {};
 
   it("Measure binary size", async () => {
+    const output = execSync("cargo metadata --no-deps --format-version=1", {
+      encoding: "utf8",
+    });
+    const metadata = JSON.parse(output);
     const stat = await fs.stat(
-      path.join("target", "deploy", `${IDL.metadata.name}.so`)
+      path.join(metadata.target_directory, "deploy", `${IDL.metadata.name}.so`)
     );
     binarySize[IDL.metadata.name] = stat.size;
   });